requirements.txt
RPi.GPIO==0.7.1
keyboard==0.13.5 